RestaurantSigloXXI/requirements.txt

djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
